pyproject.toml
[build-system]
requires = ["setuptools>=62", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stereogum-automator"
version = "0.1.0"
description = "Scrape Stereogum new music and sync to Apple Music playlist"
authors = [{ name = "" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer[all]>=0.12",
  "httpx>=0.27",
  "beautifulsoup4>=4.12",
  "lxml>=5.2",
  "python-dotenv>=1.0",
  "PyJWT[crypto]>=2.8",
  "pydantic>=2.7",
  "feedparser>=6.0",
]

[project.scripts]
stereogum = "stereogum_automator.cli:app"

[tool.setuptools]
packages = ["stereogum_automator"]

